backend/server-py/app/requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
langsmith
numpy<2.0
mediapipe==0.10.21
opencv-python
matplotlib
pytesseract
Pillow
fastapi
uvicorn[standard]
google-genai
google-cloud-aiplatform
aiofiles
asyncio
openai
scikit-image
langchain_openai
langgraph
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
python-dotenv>=1.0.0